Platform support for Windows Store and ERA
This release introduces the ability to read and deduct sparks.
It also includes support for Windows Store and ERA platforms. The
support for these platforms is in "Preview".
If you find issues, let us know by filing issues in the issues section.
This release also fixes issue:
#18 - Getting Cost of Buttons Always Returns 0
Certain properties on the Participant were not returning correct values
(for instance, LastInputAt)
Breaking Change
The InteractivityManager prefab now persists across scenes (it's marked
DontDestroyOnLoad). This was done to address issues on scene transitions
or when there was more than one scene loaded at once.
Rather than having an instance of the prefab in each scene, you will
want to have one instance of the InteractivityManager prefab.